# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.TipeSelectResponse:
    properties:
      code:
        type: string
      id:
        type: integer
      label:
        type: string
    type: object
...
  httputil.UserData:
    properties:
      id:
        type: integer
      name:
        type: string
      username:
        type: string
    type: object
...
  neosim_go_internal_modules_kepegawaian_identifier_dto.CreateTipeRequest:
    properties:
      code:
        maxLength: 100
        minLength: 1
        type: string
      description:
        maxLength: 500
        type: string
      fhir_code:
        maxLength: 500
        type: string
      fhir_system:
        maxLength: 255
        type: string
      has_expiry:
        type: boolean
      is_nakes:
        type: boolean
      is_required:
        type: boolean
      label:
        maxLength: 255
        minLength: 1
        type: string
      penerbit:
        maxLength: 255
        type: string
      point:
        minimum: 0
        type: number
    required:
    - code
    - label
    type: object
  neosim_go_internal_modules_kepegawaian_identifier_dto.TipeResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/httputil.UserData'
      description:
        type: string
      fhir_code:
        type: string
      fhir_system:
        type: string
      has_expiry:
        type: boolean
      id:
        type: integer
      is_nakes:
        type: boolean
      is_required:
        type: boolean
      label:
        type: string
      penerbit:
        type: string
      point:
        type: number
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/httputil.UserData'
    type: object
...
  neosim_go_internal_modules_kepegawaian_identifier_dto.UpdateTipeRequest:
    properties:
      code:
        maxLength: 100
        minLength: 1
        type: string
      description:
        maxLength: 500
        type: string
      fhir_code:
        maxLength: 500
        type: string
      fhir_system:
        maxLength: 255
        type: string
      has_expiry:
        type: boolean
      is_nakes:
        type: boolean
      is_required:
        type: boolean
      label:
        maxLength: 255
        minLength: 1
        type: string
      penerbit:
        maxLength: 255
        type: string
      point:
        minimum: 0
        type: number
    type: object
...
  response.MyGoResponse:
    description: Standard JSON response
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
host: localhost:1323
info:
  contact:
    email: support@neosim.id
    name: Neosim Support
  description: REST API untuk Neosim — sistem manajemen informasi
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Neosim API V1
  version: "1.0"
paths:
...
  /kepegawaian/identifier/tipes:
    get:
      consumes:
      - application/json
      description: Get paginated list of Tipe with optional filters
      parameters:
      - description: Search by label or code
        in: query
        name: search
        type: string
      - description: Filter by code
        in: query
        name: code
        type: string
      - description: Filter by label
        in: query
        name: label
        type: string
      - description: Filter by is_nakes
        in: query
        name: is_nakes
        type: boolean
      - description: Filter by is_required
        in: query
        name: is_required
        type: boolean
      - description: Sort column (code, label,is_nakes,is_required, created_at, updated_at)
        enum:
        - code
        - label
        - is_nakes
        - is_required
        - created_at
        - updated_at
        in: query
        name: sort_by
        type: string
      - description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.MyGoResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/neosim_go_internal_modules_kepegawaian_identifier_dto.TipeResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get list of Tipe
      tags:
      - kepegawaian/identifier/tipe
    post:
      consumes:
      - application/json
      description: Create New Tipe
      parameters:
      - description: Create Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/neosim_go_internal_modules_kepegawaian_identifier_dto.CreateTipeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.MyGoResponse'
            - properties:
                data:
                  $ref: '#/definitions/neosim_go_internal_modules_kepegawaian_identifier_dto.TipeResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Create Tipe
      tags:
      - kepegawaian/identifier/tipe
  /kepegawaian/identifier/tipes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Tipe by :id
      parameters:
      - description: Tipe ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MyGoResponse'
      security:
      - BearerAuth: []
      summary: Delete Tipe
      tags:
      - kepegawaian/identifier/tipe
    get:
      consumes:
      - application/json
      description: Get Tipe by :id
      parameters:
      - description: Tipe ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.MyGoResponse'
            - properties:
                data:
                  $ref: '#/definitions/neosim_go_internal_modules_kepegawaian_identifier_dto.TipeResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get Tipe
      tags:
      - kepegawaian/identifier/tipe
    put:
      consumes:
      - application/json
      description: Update Tipe by :id
      parameters:
      - description: Tipe ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/neosim_go_internal_modules_kepegawaian_identifier_dto.UpdateTipeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.MyGoResponse'
            - properties:
                data:
                  $ref: '#/definitions/neosim_go_internal_modules_kepegawaian_identifier_dto.TipeResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Update Tipe
      tags:
      - kepegawaian/identifier/tipe
  /kepegawaian/identifier/tipes/select:
    get:
      consumes:
      - application/json
      description: Get list of Tipe for select with optional search
      parameters:
      - description: Search by label or code
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.MyGoResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.TipeSelectResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get list of Tipe for select
      tags:
      - kepegawaian/identifier/tipe
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"